Prerelease Version for Israel
GPIO inputs are working and the info tab is working. Code is contained
in the project for the rest of the Micronet Hardware commands but it is
not used because the MT5 doesn't use a lot of the commands. Outputs are
currently not working on the app. Info tab needs to check the whether
the device is docked or undocked so it can auto update the ui when the
mcu and fpga are available.